Moved generateRandomIndices out to separate module
